Rename the trajectory planner and start calculating progress
Argh! I was really just pulling a Homer here. I was shadowing the
actual trajectory planner and of course things went off the rails when
I started adding variables. So, I renamed the thing and saw that I was
accidentally pulling in the system one. Whoops! The rename keeps me
honest and removes this possibility.

On the positive side it seems that I managed to find a way to map
progress (sort of), more work probably needs to be done.