Adjust the acceleration progress so to matches closer to the sample
This isn't quite right yet, but it seems that I at least get part of
the curve up, I don't get the curve down unfortunately. So, that must
be examined. We did get rid of the wildness of wandering around when
just turn though.

Also recalculate the change in progress correctly at least for speed
up.