Some code clean up and fix on sampling min & max
Fix our calculation of the max and min speed at the start of create
trajectory. We only need to move the acceleration up or down. Don't
try to outsmart the function. For the moment, we copy and paste the
velocity functions. But I'm OK with that.

We never really need to calculate the progress at the start since we
are constantly needing to get a new progress and the function seems
"fast enough". Plus we are always calling progress for speed with the
same arguments. So, just get rid of them and cache the total
acceleration time since the maximum speed and acceleration time rarely
changes.

Also added several different debug items.