Use std::swap for swapping trajectories.
I don't know why the trajectory stuff is being done the other
way. Conceivably it's for an optimization so we don't allocate
them. But, looking at the class, I don't see how we are really saving
here. Instead, use std::swap, which gives us moves semantics and
should be as efficent as this was imagined to be.