Update the GUI to handle changes to the move_base script
I ported the script to work with Patrick's changes
and work with the torso (this addresses Issue #2).

To make my life simpler, I also created a fetch_controller class that
handles all of the process stuff. This stops that stuff from bleeding
into window GUI code and keeps it a bit cleaner. Since I'm adding a
lot of complexity with other items, I think this is necessary.