Add some basic reading and writing of the die pool.
This is more of a guard against the fact that if we start and stop the
GUI multiple times we may get in the extremely unlikely case that we
happen to regenerate the same die roll at the end. So, we generate the
pool once, save it, write it out, and read it back when we start again.

At the moment, I'm just dumping the pool in the current working
directory. That is not going to work long term. This of course means
that I need a parameter, but I needed to do that anyway for the logging.