Make the code more general.
This is just to make my like a little more bearable. The idea is to
make it easy to grab the variable names in whatever function I'm using
and to leverage what I know about functional languages to avoid
repeating myself. I could go a bit more crazy here, but for the moment
this is OK. For example, the median stuff is not that interesting to
me, but it helped me to see that things are working as is.