Fix variable name make a summary functions
Anthro not Antro.

Also added a nice function for putting the
statistics into a table. It looks pretty good, and it seems that I'm
doing the correct thing with list programming. It still needs another
function to curry so that I can put in through a pipe, but I think
it's more or less there and I added the basics in a comment.