Go full on tidyverse.
Since I was trying to use the ggplot2 boxplot, it seems I need to have
tidy data. This took much longer than I thought, but I think the
result is fine.

With tidy data, I hope that it will actually be easier to calculate
all the stuff. I kept the "raw" version around, so that I don't break
all my code.

Here's hoping that everything from now on is downhill.